chore: rename dialog-ucan to dialog-ucan-core
The UCAN token implementation is being separated from the capability
protocol layer. This renames the low-level UCAN spec implementation
to dialog-ucan-core, freeing the dialog-ucan name for the capability
protocol integration (coming in a later PR).

Pure rename: no code changes, just crate name and identifier updates.
