+++ Cargo.toml
@@ -29,7 +29,7 @@ members = [
     "rust/dialog-query",
     "rust/dialog-search-tree",
     "rust/dialog-storage",
-    "rust/dialog-ucan",
+    "rust/dialog-ucan-core",
     "rust/dialog-varsig",
     "rust/dialog-blobs",
 ]
@@ -139,7 +139,7 @@ tokio-util = { version = "0.7", features = ["io"] }
 tower = "0.5"
 tower-http = "0.6"
 tracing = "0.1"
-dialog-ucan = { path = "./rust/dialog-ucan" }
+dialog-ucan-core = { path = "./rust/dialog-ucan-core" }
 dialog-varsig = { path = "./rust/dialog-varsig" }
 ucan = { git = "https://github.com/tonk-labs/rs-ucan.git", branch = "main", package = "ucan" }
 varsig = { git = "https://github.com/tonk-labs/rs-ucan.git", branch = "main", package = "varsig" }
