+++ rust/dialog-ucan-core/src/command.rs
